+++ web/modules/custom/qs_activity/src/Form/EventAddForm.php
@@ -427,10 +427,14 @@ class EventAddForm extends FormBasic {
       $repeat_period = 'W';
     }
 
+    // An array collection of created events.
+    $events = [];
+
     // Create one event once, then repeat the event if necessary.
     do {
       // Create the new event.
       $event = $this->eventManager->create($activity, $start_at, $end_at, $data);
+      $events[] = $event;
 
       // Get the current user activitiy's privilege to this event.
       $privileges_by_events = $this->badgeManager->getPrivilegesByEvents([$event]);
@@ -456,17 +460,20 @@ class EventAddForm extends FormBasic {
     switch ($trigger['#name']) {
       case 'save_and_repeat_weekly':
         $this->messenger()->addMessage($this->t('qs_activity.events.form.add.weekly.success @repeat', ['@repeat' => $repeat]));
-        $form_state->setRedirect('entity.node.canonical', ['node' => $activity->id()]);
 
         break;
 
       case 'save':
       default:
         $this->messenger()->addMessage($this->t('qs_activity.events.form.add.success'));
-        $form_state->setRedirect('entity.node.canonical', ['node' => $activity->id()], ['fragment' => 'card' . $event->id()]);
 
         break;
     }
+
+    // Get the first event to point on redirect.
+    $first_event = reset($events);
+
+    $form_state->setRedirect('entity.node.canonical', ['node' => $activity->id()], ['fragment' => 'card' . $first_event->id()]);
   }
 
   /**