+++ CHANGELOG.md
@@ -14,6 +14,7 @@
  - add QuillJS link capability - #849
  - fix Excel future events export date timezone - #802
  - fix navigation button next months on Photos by Month, Icon position left instead of right - #759
+ - on multiple events creation (#803) add the URL fragment to the first created event - #854
 
 ## 1.6.1 (2021-04-23)
  - update drupal/core (8.9.13 => 8.9.14)